+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add catalog and kubernetes labels and notes annotation.
+
 ## [3.4.1] - 2020-10-29
 
 ### Added